Extract Apple Publisher model
