+++ app/models/apple/show.rb
@@ -109,32 +109,6 @@ class Apple::Show
     self.class.get_episodes(api, external_id)
   end
 
-  def get_episode_asset_container_metadata
-    get_episodes.map do |ep|
-      vendor_id = ep['attributes']['appleHostedAudioAssetVendorId']
-
-      {
-        apple_episode_id: ep['id'],
-        audio_asset_vendor_id: vendor_id,
-        podcast_containers_url: episode_podcast_container_url(vendor_id)
-      }
-    end
-  end
-
-  def episode_podcast_container_url(vendor_id)
-      api.join_url('podcastContainers?filter[vendorId]=' + vendor_id).to_s
-  end
-
-  def get_podcast_containers
-
-    resp =
-      api.bridge_get('podcastContainers', get_episode_asset_container_metadata)
-
-    binding.pry
-
-    api.unwrap_response(resp)
-  end
-
   def self.get_show(api, show_id)
     resp = api.get("shows/#{show_id}")
